cloud_functions/climateiq_data_export_cf/requirements.txt
black==24.4.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
flake8==7.0.0
Flask==3.0.3
functions-framework==3.5.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathlib==1.0.1
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.1.1
requests==2.31.0
rsa==4.9
typing==3.7.4.3
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.2
